requirements.txt
openai
streamlit
pypdf2
python-docx

reportlab
